+++ src/components/Button/index.ts
@@ -1,4 +1,4 @@
-// Button from copy-ui @ 2025-10-30.
+// Button from copy-ui @ 2025-11-15.
 
 export type { ButtonProps } from "./button";
 export { Button } from "./button";